pyproject.toml
[tool.poetry]
name = "auto_gram"
version = "0.1.0"
description = ""
authors = ["fagja <hey.shunya19@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.20.0"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
